backend/requirements.txt
flask
flask-cors
groq
langchain
langchain-community
langchain-openai
faiss-cpu
pandas
numpy
scikit-learn
nltk
unidecode
tqdm
streamlit
firebase-admin
python-dotenv
requests
flask-restful
flask-cors
sentence_transformers
rapidfuzz
pickle
matcher